MODE-1558 - Created a new modeshape-webdav-war module which should be use for integration testing the standard webdav implementation, provided via the modeshape-webdav module. Also, updated the existing webdav integration tests to use the new module.
