+++ web/modeshape-webdav-war/src/main/resources/log4j.properties
@@ -1,0 +1,11 @@
+log4j.rootLogger = INFO, stdout
+
+log4j.category.org.apache=INFO
+log4j.category.org.mortbay.jetty.security=ERROR
+log4j.category.org.slf4j.impl.JCLLoggerAdapter=INFO
+log4j.category.org.springframework=INFO
+
+log4j.appender.stdout = org.apache.log4j.ConsoleAppender
+log4j.appender.stdout.Target   = System.out
+log4j.appender.stdout.layout = org.apache.log4j.PatternLayout
+log4j.appender.stdout.layout.ConversionPattern = [%-5p] [%C] : %m%n [%F:%L]
