+++ web/modeshape-webdav-war/src/main/webapp/META-INF/MANIFEST.MF
@@ -1,0 +1,1 @@
+Manifest-Version: 1.0
