+++ web/modeshape-webdav-war/src/test/resources/textfile.txt
@@ -1,0 +1,1 @@
+Some text content to be created in WebDav.
